+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/report/compare/release-3.4.0...master
+[Unreleased]: https://github.com/atlassian/report/compare/release-3.5.0...master
+
+## [3.5.0] - 2019-07-05
+[3.5.0]: https://github.com/atlassian/report/compare/release-3.4.0...release-3.5.0
 
 ### Fixed
 - Inform about over-cropping action metric timelines. Fix [JPERF-152].